+++ pkg/logging/global.go
@@ -1,7 +1,7 @@
 //
 // DISCLAIMER
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -21,18 +21,27 @@
 package logging
 
 import (
+	"io"
 	"os"
 	"time"
 
 	"github.com/rs/zerolog"
 )
 
-var global = NewFactory(zerolog.New(zerolog.ConsoleWriter{
-	Out:        os.Stdout,
-	TimeFormat: time.RFC3339Nano,
-	NoColor:    true,
-}).With().Timestamp().Logger())
+var global = SetGlobal(true)
 
 func Global() Factory {
 	return global
+}
+
+func SetGlobal(isPretty bool) Factory {
+	var w io.Writer = os.Stderr
+	if isPretty {
+		w = zerolog.ConsoleWriter{
+			Out:        os.Stdout,
+			TimeFormat: time.RFC3339Nano,
+			NoColor:    true,
+		}
+	}
+	return NewFactory(zerolog.New(w).With().Timestamp().Logger())
 }