+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - (Feature) Optional ResignLeadership Action
 - (Feature) Improve CRD Management and deprecate CRD Chart
 - (Bugfix) Fix invalid Timeout calculation in case of ActionList
+- (Feature) Optional JSON logger format
 
 ## [1.2.24](https://github.com/arangodb/kube-arangodb/tree/1.2.24) (2023-01-25)
 - (Bugfix) Fix deployment creation on ARM64